src/backend/requirements.txt
Flask
Flask-PyMongo
python-dotenv
pyjwt
twilio